gc: Nicer errors before miscompiling.
This fixes issue 2444.

A big cleanup of all 31/32bit size boundaries i'll leave for another cl though.  (see also issue 1700).

R=rsc
CC=golang-dev
https://golang.org/cl/5484058

+++ src/cmd/6g/cgen.c
@@ -1023,13 +1023,13 @@ stkof(Node *n)
  *	memmove(&ns, &n, w);
  */
 void
-sgen(Node *n, Node *ns, int32 w)
+sgen(Node *n, Node *ns, int64 w)
 {
 	Node nodl, nodr, oldl, oldr, cx, oldcx, tmp;
 	int32 c, q, odst, osrc;
 
 	if(debug['g']) {
-		print("\nsgen w=%d\n", w);
+		print("\nsgen w=%lld\n", w);
 		dump("r", n);
 		dump("res", ns);
 	}
@@ -1038,7 +1038,7 @@ sgen(Node *n, Node *ns, int32 w)
 		fatal("sgen UINF");
 
 	if(w < 0)
-		fatal("sgen copy %d", w);
+		fatal("sgen copy %lld", w);
 
 	if(w == 16)
 		if(componentgen(n, ns))
